# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Controlled Verify V1 product configuration
  - name: account
    description: Customer account and dashboard operations
  - name: transactions
    description: Guest and authenticated transaction operations
  - name: admin
    description: Protected administration operations
  - name: contact
    description: Public support contact and protected message review
  - name: webhooks
    description: Provider and authentication webhook structure
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List configured Verify products
      responses:
        "200":
          description: Configured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{slug}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get one configured Verify product
      parameters:
        - $ref: "#/components/parameters/ProductSlug"
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /me:
    get:
      operationId: getCurrentUser
      tags: [account]
      summary: Get the current authenticated user
      security:
        - clerkSession: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [account]
      summary: Get the current customer dashboard summary
      security:
        - clerkSession: []
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions for the current customer
      security:
        - clerkSession: []
      responses:
        "200":
          description: Customer transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a guest or authenticated transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Invalid transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /transactions/{reference}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction by DepositSafe reference
      parameters:
        - $ref: "#/components/parameters/TransactionReference"
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /companies-house/search:
    get:
      operationId: searchCompaniesHouseCompanies
      tags: [companies-house]
      summary: Search Companies House companies
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 120
      responses:
        "200":
          description: Matching public company records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySearchResponse"
        "503":
          description: Companies House is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /transactions/{reference}/companies-house/check:
    post:
      operationId: runCompaniesHouseCheck
      tags: [companies-house]
      summary: Check a selected Companies House company for a transaction
      security:
        - clerkSession: []
        - guestCapability: []
      parameters:
        - $ref: "#/components/parameters/TransactionReference"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyCheckInput"
      responses:
        "200":
          description: Stored Companies House result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyCheckResult"
        "403":
          description: Valid transaction access is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /transactions/{reference}/companies-house/result:
    get:
      operationId: getCompaniesHouseResult
      tags: [companies-house]
      summary: Get the Companies House result linked to a transaction
      security:
        - clerkSession: []
        - guestCapability: []
      parameters:
        - $ref: "#/components/parameters/TransactionReference"
      responses:
        "200":
          description: Companies House result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyCheckResult"
        "404":
          description: No Companies House result exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /transactions/{reference}/payments/checkout:
    post:
      operationId: createStripeCheckoutSession
      tags: [payments]
      summary: Create a Stripe Checkout session for an existing transaction
      security:
        - clerkSession: []
        - guestCapability: []
      parameters:
        - $ref: "#/components/parameters/TransactionReference"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StripeCheckoutInput"
      responses:
        "201":
          description: Stripe Checkout session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StripeCheckoutResponse"
        "409":
          description: Transaction is not eligible for checkout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/providers/stripe/catalog-sync:
    post:
      operationId: syncStripeCatalog
      tags: [admin]
      summary: Map existing Stripe products/prices to the DepositSafe catalogue
      description: Maps the runtime environment's existing Stripe prices to locked DepositSafe products without creating or changing live Stripe objects.
      security:
        - clerkSession: []
      responses:
        "200":
          description: Stripe catalogue synchronization summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StripeCatalogSyncResponse"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/transactions:
    get:
      operationId: listAdminTransactions
      tags: [admin]
      summary: List all transactions for authorised administrators
      security:
        - clerkSession: []
      responses:
        "200":
          description: All transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/contact-messages:
    get:
      operationId: listAdminContactMessages
      tags: [admin]
      summary: List recent public contact form submissions
      security:
        - clerkSession: []
      responses:
        "200":
          description: Recent contact messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactMessage"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /contact-messages:
    post:
      operationId: submitContactMessage
      tags: [contact]
      summary: Submit a public support enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactMessageInput"
      responses:
        "201":
          description: Enquiry accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmissionResponse"
        "400":
          description: Invalid enquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Enquiry storage is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many enquiries in a short period
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /webhooks/clerk:
    post:
      operationId: receiveClerkWebhook
      tags: [webhooks]
      summary: Reserved Clerk webhook endpoint
      description: Build 01 reserves this endpoint for future user synchronization. It does not call external providers.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookInput"
      responses:
        "202":
          description: Accepted for future processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookAccepted"
  /webhooks/stripe:
    post:
      operationId: receiveStripeWebhook
      tags: [webhooks]
      summary: Receive a signed Stripe event
      description: Stripe-Signature is verified against the raw request bytes using the managed webhook signing secret before the event is processed.
      parameters:
        - name: Stripe-Signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StripeWebhookEvent"
      responses:
        "200":
          description: Event accepted or already processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderWebhookResponse"
        "400":
          description: Invalid signature or payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Webhook signing secret is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    clerkSession:
      type: http
      scheme: bearer
      description: Clerk session cookie is used by the browser; bearer notation documents the protected contract.
    guestCapability:
      type: apiKey
      in: header
      name: X-Guest-Capability
      description: One-time issued random capability for a guest transaction; stored only as a hash server-side.
  parameters:
    ProductSlug:
      name: slug
      in: path
      required: true
      schema:
        type: string
    TransactionReference:
      name: reference
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Product:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        pricePence:
          type: integer
        price:
          type: string
        provider:
          type: string
        participantMode:
          type: string
          enum: [single, multiple]
        active:
          type: boolean
      required: [slug, name, pricePence, price, provider, participantMode, active]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [customer, admin]
      required: [id, email, displayName, role]
    ParticipantInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
      required: [name, email]
    Participant:
      allOf:
        - $ref: "#/components/schemas/ParticipantInput"
        - type: object
          properties:
            id:
              type: string
          required: [id]
    TransactionInput:
      type: object
      properties:
        productSlug:
          type: string
        email:
          type: string
        participant:
          $ref: "#/components/schemas/ParticipantInput"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ParticipantInput"
      required: [productSlug, email]
    Transaction:
      type: object
      properties:
        reference:
          type: string
        product:
          $ref: "#/components/schemas/Product"
        email:
          type: string
        status:
          $ref: "#/components/schemas/TransactionStatus"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/Participant"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isGuest:
          type: boolean
        guestCapability:
          type: string
          nullable: true
      required: [reference, product, email, status, participants, createdAt, updatedAt, isGuest]
    CompanySearchResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CompanySearchResult"
      required: [items]
    CompanySearchResult:
      type: object
      properties:
        companyNumber:
          type: string
        name:
          type: string
        status:
          type: string
        dateOfCreation:
          type: string
          nullable: true
        addressSnippet:
          type: string
          nullable: true
      required: [companyNumber, name, status]
    CompanyCheckInput:
      type: object
      properties:
        companyNumber:
          type: string
          minLength: 1
          maxLength: 8
      required: [companyNumber]
    CompanyCheckResult:
      type: object
      properties:
        verificationId:
          type: string
        transactionReference:
          type: string
        companyNumber:
          type: string
        companyName:
          type: string
        companyStatus:
          type: string
        outcome:
          type: string
        transactionStatus:
          $ref: "#/components/schemas/TransactionStatus"
        checkedAt:
          type: string
          format: date-time
      required: [verificationId, transactionReference, companyNumber, companyName, companyStatus, outcome, transactionStatus, checkedAt]
    StripeCheckoutInput:
      type: object
      properties:
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
        successUrl:
          type: string
          format: uri
        cancelUrl:
          type: string
          format: uri
      required: [idempotencyKey, successUrl, cancelUrl]
    StripeCheckoutResponse:
      type: object
      properties:
        checkoutSessionReference:
          type: string
        checkoutUrl:
          type: string
          format: uri
        paymentStatus:
          type: string
          enum: [pending, paid, failed]
        amountPence:
          type: integer
      required: [checkoutSessionReference, checkoutUrl, paymentStatus, amountPence]
    StripeCatalogSyncResponse:
      type: object
      properties:
        productsSynced:
          type: integer
        pricesSynced:
          type: integer
      required: [productsSynced, pricesSynced]
    StripeWebhookEvent:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        created:
          type: integer
        data:
          type: object
          additionalProperties: true
      required: [id, type, created, data]
      additionalProperties: true
    ProviderWebhookResponse:
      type: object
      properties:
        accepted:
          type: boolean
        duplicate:
          type: boolean
        stale:
          type: boolean
      required: [accepted, duplicate, stale]
    ContactMessageInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        email:
          type: string
          format: email
          maxLength: 254
        topic:
          type: string
          enum: [general, order, privacy, other]
        message:
          type: string
          minLength: 10
          maxLength: 4000
        website:
          type: string
          maxLength: 200
          description: Hidden anti-spam field; leave empty.
      required: [name, email, topic, message]
    ContactSubmissionResponse:
      type: object
      properties:
        accepted:
          type: boolean
        reference:
          type: string
      required: [accepted, reference]
    ContactMessage:
      type: object
      properties:
        reference:
          type: string
        name:
          type: string
        email:
          type: string
        topic:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [reference, name, email, topic, message, createdAt]
    TransactionStatus:
      type: string
      enum:
        - STARTED
        - PAYMENT_PENDING
        - PAID
        - VERIFICATION_PENDING
        - VERIFICATION_IN_PROGRESS
        - VERIFICATION_COMPLETED
        - RESULT_GENERATED
        - DELIVERED
        - PAYMENT_FAILED
        - AWAITING_PARTICIPANT
        - EXPIRED
        - VERIFICATION_FAILED
        - MANUAL_ATTENTION
    DashboardSummary:
      type: object
      properties:
        totalTransactions:
          type: integer
        activeTransactions:
          type: integer
        completedTransactions:
          type: integer
        recentTransactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
      required: [totalTransactions, activeTransactions, completedTransactions, recentTransactions]
    WebhookInput:
      type: object
      additionalProperties: true
    WebhookAccepted:
      type: object
      properties:
        accepted:
          type: boolean
      required: [accepted]